wip - tf variable a common place: need to have sim-link - local end init scripts moved in this project
