+++ changelog/0.5.md
@@ -4,7 +4,7 @@
 - dynamodb tables
 - kms master key
 - s3 document repository
-- default mail templates (only for demo \)
+- default mail templates (only for demo)
 ## :heart: Contributor
 
 - [@mrflick72](https://github.com/mrFlick72)
