+++ changelog/0.6.md
@@ -1,0 +1,6 @@
+# VAuthenticator provisioner environment infrastructure 0.6
+
+## :star: New Features
+## :heart: Contributor
+
+- [@mrflick72](https://github.com/mrFlick72)
