+++ local-environment/init.sh
@@ -1,0 +1,8 @@
+source .env
+
+cd ../scripts
+ls -la
+pip3 install -r requirements.txt
+
+python3 key_setup.py $MASTER_KEY $TABLES_SUFFIX
+#python3 setup.py admin@email.com $MASTER_KEY
