--- terraform/iam/variable.tf
@@ -1,12 +1,0 @@
-# tags variables
-variable "common_resource_tags" {
-  type = map(string)
-}
-
-# iam variables
-variable "username" {
-  type = string
-}
-variable "path" {
-  type = string
-}
